0012788: allow acl for all folder nodes (update script)
* migrates old container acl to node acl

https://forge.tine20.org/view.php?id=12788

Reviewed-on: http://gerrit.tine20.com/customers/4522
Tested-by: Jenkins CI (http://ci.tine20.com/)

+++ tine20/Tinebase/Container.php
@@ -1138,7 +1138,12 @@ class Tinebase_Container extends Tinebase_Backend_Sql_Abstract implements Tineba
                 $acl = $controller->doContainerACLChecks(FALSE);
             }
 
-            $controller->deleteContainerContents($container, $_ignoreAcl);
+            if (method_exists($controller, 'deleteContainerContents')) {
+                $controller->deleteContainerContents($container, $_ignoreAcl);
+            } else {
+                Tinebase_Core::getLogger()->notice(__METHOD__ . '::' . __LINE__
+                    . ' No deleteContainerContents defined in controller ' . get_class($controller));
+            }
 
             if ($_ignoreAcl === TRUE && method_exists($controller, 'doContainerACLChecks')) {
                 $controller->doContainerACLChecks($acl);
